refactor(kwin-effect): extract daemon-bringup to plasmazoneseffect/daemon_bringup.cpp
Pulls 9 daemon-readiness, settings-sync, and navigation-signal-wiring
methods out of plasmazoneseffect.cpp: slotDaemonReady,
continueDaemonReadySetup, processDaemonReadyWindowState,
slotSettingsChanged, loadSettingAsync (template), loadCachedSettings,
anyLocalTriggerHeld, detectActivationAndGrab, connectNavigationSignals.
Also moves the TriggerModifierField/TriggerMouseButtonField anon-namespace
constants to the new TU. Drops the main file from 3792 to 3110 lines.
